add support for the photo archive.
To support the photo gallery we add support for bootstrap and use the album
tag. This lays out a set of photos in a grid, with the ability to caption
them.

We also add a new data file for the photographs metadata.

And finally, we add a first set of photos from Antarctica.

Would be nice to figure out how to use collections and generate the index
files automatically.